Updates - Cleanup Formatting

+++ controllers/gallery.php
@@ -27,7 +27,7 @@ function getComments($id_image)
 	try {
 		$conn = dbConnect();
 		$stmt = $conn->prepare("SELECT * FROM comments WHERE id_image = :id_image ORDER BY id_comment ASC");
-		$stmt->bindParam(':id_image', $id_image);
+		$stmt->bindParam(':id_image', $id_image, PDO::PARAM_INT);
 		$stmt->execute();
 		return $stmt->fetchAll();
 	} catch (PDOException $e) {
@@ -44,29 +44,37 @@ if (!empty($_POST['deleteButton']) && isset($_POST['id_image'])) {
 
 function deleteImage($id_image)
 {
-	try {
-		$conn = dbConnect();
-		$stmt = $conn->prepare("SELECT * FROM images WHERE id_image = :id_image");
-		$stmt->bindParam(':id_image', $id_image);
-		$stmt->execute();
-		$image = $stmt->fetch();
-		if ($image['id_user'] == $_SESSION['id_user']) {
-			$imagePath = $image['image_name'];
-			unlink("../public/uploads/" . $imagePath);
-			$stmt = $conn->prepare("DELETE FROM images WHERE id_image = :id_image");
-			$stmt->bindParam(':id_image', $id_image);
-			$stmt->execute();
-			$stmt = $conn->prepare("DELETE FROM comments WHERE id_image = :id_image");
-			$stmt->bindParam(':id_image', $id_image);
-			$stmt->execute();
-			$stmt = $conn->prepare("DELETE FROM likes WHERE id_image = :id_image");
-			$stmt->bindParam(':id_image', $id_image);
-			$stmt->execute();
+	if (isset($_SESSION['check']) && !empty($_SESSION['check'])) {
+		if (($_SESSION['check']) != hash('ripemd128', $_SERVER['REMOTE_ADDR'] . $_SERVER['HTTP_USER_AGENT'])) {
+			session_destroy();
+			header('Location: ../sources/login.html.php?error=Session expired!');
+		} else {
+			$_SESSION['check'] = hash('ripemd128', $_SERVER['REMOTE_ADDR'] . $_SERVER['HTTP_USER_AGENT']);
+			try {
+				$conn = dbConnect();
+				$stmt = $conn->prepare("SELECT * FROM images WHERE id_image = :id_image");
+				$stmt->bindParam(':id_image', $id_image, PDO::PARAM_INT);
+				$stmt->execute();
+				$image = $stmt->fetch();
+				if ($image['id_user'] == $_SESSION['id_user']) {
+					$imagePath = $image['image_name'];
+					unlink("../public/uploads/" . $imagePath);
+					$stmt = $conn->prepare("DELETE FROM images WHERE id_image = :id_image");
+					$stmt->bindParam(':id_image', $id_image, PDO::PARAM_INT);
+					$stmt->execute();
+					$stmt = $conn->prepare("DELETE FROM comments WHERE id_image = :id_image");
+					$stmt->bindParam(':id_image', $id_image, PDO::PARAM_INT);
+					$stmt->execute();
+					$stmt = $conn->prepare("DELETE FROM likes WHERE id_image = :id_image");
+					$stmt->bindParam(':id_image', $id_image, PDO::PARAM_INT);
+					$stmt->execute();
 
-			header('Location: ../sources/home.html.php?success=Image deleted!');
+					header('Location: ../sources/home.html.php?success=Image deleted!');
+				}
+			} catch (PDOException $e) {
+				echo $e->getMessage() . " in " . $e->getFile() . ":" . $e->getLine();
+				exit();
+			}
 		}
-	} catch (PDOException $e) {
-		echo $e->getMessage() . " in " . $e->getFile() . ":" . $e->getLine();
-		exit();
 	}
 }
