+++ controllers/likes.php
@@ -2,6 +2,7 @@
 
 session_start();
 
+
 include_once '../config/dbConnect.php';
 
 date_default_timezone_set('Europe/Helsinki');
@@ -20,8 +21,8 @@ function checkLikes($id_image, $id_user)
 	try {
 		$conn = dbConnect();
 		$stmt = $conn->prepare("SELECT * FROM likes WHERE id_image = :id_image AND id_user = :id_user");
-		$stmt->bindParam(':id_image', $id_image);
-		$stmt->bindParam(':id_user', $_SESSION['id_user']);
+		$stmt->bindParam(':id_image', $id_image, PDO::PARAM_INT);
+		$stmt->bindParam(':id_user', $_SESSION['id_user'], PDO::PARAM_INT);
 		$stmt->execute();
 		$likes = $stmt->fetchAll();
 		return $likes;
@@ -34,17 +35,25 @@ function checkLikes($id_image, $id_user)
 
 function postLikes($id_image, $id_user)
 {
-	try {
-		$conn = dbConnect();
-		$stmt = $conn->prepare("INSERT INTO likes (id_image, id_user, liked) VALUES (:id_image, :id_user, 1)");
-		$stmt->bindParam(':id_image', $id_image);
-		$stmt->bindParam(':id_user', $_SESSION['id_user']);
-		$stmt->execute();
-		header("Location: ../sources/home.html.php");
-		exit();
-	} catch (PDOException $e) {
-		echo $e->getMessage() . " in " . $e->getFile() . ":" . $e->getLine();
-		exit();
+	if (isset($_SESSION['check']) && !empty($_SESSION['check'])) {
+		if (($_SESSION['check']) != hash('ripemd128', $_SERVER['REMOTE_ADDR'] . $_SERVER['HTTP_USER_AGENT'])) {
+			session_destroy();
+			header('Location: ../sources/login.html.php?error=Session expired!');
+		} else {
+			$_SESSION['check'] = hash('ripemd128', $_SERVER['REMOTE_ADDR'] . $_SERVER['HTTP_USER_AGENT']);
+			try {
+				$conn = dbConnect();
+				$stmt = $conn->prepare("INSERT INTO likes (id_image, id_user, liked) VALUES (:id_image, :id_user, 1)");
+				$stmt->bindParam(':id_image', $id_image, PDO::PARAM_INT);
+				$stmt->bindParam(':id_user', $_SESSION['id_user'], PDO::PARAM_INT);
+				$stmt->execute();
+				header("Location: ../sources/home.html.php");
+				exit();
+			} catch (PDOException $e) {
+				echo $e->getMessage() . " in " . $e->getFile() . ":" . $e->getLine();
+				exit();
+			}
+		}
 	}
 }
 
@@ -54,7 +63,7 @@ function countLikes($id_image)
 	try {
 		$conn = dbConnect();
 		$stmt = $conn->prepare("SELECT COUNT(*) FROM likes WHERE id_image = :id_image");
-		$stmt->bindParam(':id_image', $id_image);
+		$stmt->bindParam(':id_image', $id_image, PDO::PARAM_INT);
 		$stmt->execute();
 		$count = $stmt->fetchColumn();
 		return $count;
@@ -67,17 +76,25 @@ function countLikes($id_image)
 
 function deleteLikes($id_image, $id_user)
 {
-	try {
-		$conn = dbConnect();
-		$stmt = $conn->prepare("DELETE FROM likes WHERE id_image = :id_image AND id_user = :id_user");
-		$stmt->bindParam(':id_image', $id_image);
-		$stmt->bindParam(':id_user', $_SESSION['id_user']);
-		$stmt->execute();
-		header("Location: ../sources/home.html.php");
-		exit();
-	} catch (PDOException $e) {
-		echo $e->getMessage() . " in " . $e->getFile() . ":" . $e->getLine();
-		exit();
+	if (isset($_SESSION['check']) && !empty($_SESSION['check'])) {
+		if (($_SESSION['check']) != hash('ripemd128', $_SERVER['REMOTE_ADDR'] . $_SERVER['HTTP_USER_AGENT'])) {
+			session_destroy();
+			header('Location: ../sources/login.html.php?error=Session expired!');
+		} else {
+			$_SESSION['check'] = hash('ripemd128', $_SERVER['REMOTE_ADDR'] . $_SERVER['HTTP_USER_AGENT']);
+			try {
+				$conn = dbConnect();
+				$stmt = $conn->prepare("DELETE FROM likes WHERE id_image = :id_image AND id_user = :id_user");
+				$stmt->bindParam(':id_image', $id_image, PDO::PARAM_INT);
+				$stmt->bindParam(':id_user', $_SESSION['id_user'], PDO::PARAM_INT);
+				$stmt->execute();
+				header("Location: ../sources/home.html.php");
+				exit();
+			} catch (PDOException $e) {
+				echo $e->getMessage() . " in " . $e->getFile() . ":" . $e->getLine();
+				exit();
+			}
+		}
 	}
 }
